@vmg 2021-06-23
OK! Finished reading through the PR! It looks pretty good, I like the direction this is heading. I have two main concerns:

- I think the locking is not meaningful: with 3 recursive locking states (read `m`, write `m` and rw `qMu`), I can't demonstrate that the system doesn't deadlock (mostly because I'm not a very smart guy but regardless), and particularly for `qMu`, it seems like a sub-lock for `m`-write where the extra granularity couldn't possibly reduce actual lock contention. I have a hunch that removing all the locks and keeping a single `sync.Mutex` wouldn't affect performance in a measurable way.

- I think the usage of a full heap is overkill. `container/heap` in the standard library is a bit of a trainwreck that requires a lot of boilerplate to provide very poor performance (since everything goes through `interface{}` anyway). My assumption is that simply keeping the eviction queue in a slice and _sorting it lazily_ will provide the same or better performance than a heap (since fixing the heap on every refcount update is prohibitively expensive) and greatly simplify the whole client.

I've put these two H Y P O T H E S I S to the test by, huh, well, implementing and benchmarking them.

You can see my changes in https://github.com/vitessio/vitess/commit/be8fedea67fdc1e85b89e14ab957841cbf6cfdf2

And the benchmark before/after:

```
name                 old time/op  new time/op  delta
CachedConnClient-16   144µs ± 2%   143µs ± 1%   ~     (p=0.113 n=10+9)
```

Would you mind reviewing my commit and cherry picking it into this PR? I'm hoping you'll agree the code is meaningfully simpler!

That's from a perf/correctness point of view. Besides that, the other global refactorings look good to me. Changing the return for the methods into the client + `io.Closer` seems like the best way to support pooled and unpooled at the same time. :+1: