@vmg 2021-06-24
@ajm188 I've rebased my patch and fixed your suggestions: https://github.com/vitessio/vitess/commit/7bc81d816e56d3d79fa28bbd94c55ac214736492

I think we're now always re-sorting properly and I don't think I can find any actual performance regressions. Would you mind reviewing it again? I would very much prefer to see simplified locking on this connection cache!