@vmg 2021-06-25
Sounds great. Have a good time!